Ive fixed the issue with the sound not playing. The issue occured because the program was shutting down before the music had a chance.
I tried different ways to solve this, such as a loop - which just made
the music play on top of one another

I also tried using timers - but i thought there had to be a better way.

So i landed on using an input function - which to be honest i dont know
why i didnt think of that earlier.

The program works fine now.

I'm also going to add the music file to this dir for easier accesss.

Thank you.